move test views to the tests namespace
